more complete initService.sh script

+++ scripts/initService.sh
@@ -1,2 +1,5 @@
 #!/bin/bash
-ln -s /var/id-service/*.jar /etc/init.d/id-service
+if [ ! -L /etc/init.d/id-service ]
+then
+  ln -s /var/id-service/*.jar /etc/init.d/id-service
+fi
